requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
mysql-connector-python==9.1.0
pymilvus==2.5.0
spacy==3.8.3
transformers==4.47.0
sentence-transformers==3.3.1